requirements.txt
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.1
torch>=2.3.0
torchvision>=0.18.0
sentence-transformers>=3.0.1
tqdm>=4.66.4
Pillow>=10.3.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
streamlit>=1.36.0
requests>=2.32.3

